Alterada interface para transações (antigo extrato)
