+++ lib/pjbank/recebimento/base.rb
@@ -7,6 +7,14 @@ module PJBank
         @http = http
       end
 
+      def emitir(dados)
+        http.post("#{base_url_path}/:credencial/transacoes", { payload: dados })
+      end
+
+      def transacoes(params={})
+        http.get("#{base_url_path}/:credencial/transacoes", params)
+      end
+
       private
 
       def base_url_path
