+++ lib/pjbank/recebimento/boleto.rb
@@ -7,10 +7,6 @@ module PJBank
         http.post("#{base_url_path}/", { payload: dados })
       end
 
-      def emitir(dados)
-        http.post("#{base_url_path}/:credencial/transacoes", { payload: dados })
-      end
-
       def impressao(dados)
         http.post("#{base_url_path}/:credencial/transacoes/lotes", { payload: dados })
       end
