+++ lib/pjbank/recebimento/cartao.rb
@@ -11,10 +11,6 @@ module PJBank
         http.post("#{base_url_path}/:credencial/tokens", { payload: dados })
       end
 
-      def emitir(dados)
-        http.post("#{base_url_path}/:credencial/transacoes", { payload: dados })
-      end
-
       def cancelar(tid)
         http.delete("#{base_url_path}/:credencial/transacoes/#{tid}")
       end
