+++ spec/pjbank/recebimento/cartao_spec.rb
@@ -213,4 +213,23 @@ RSpec.describe PJBank::Recebimento::Cartao do
       end
     end
   end
+
+  describe "#transacoes" do
+    it "returns all the transactions an array of objects" do
+      VCR.use_cassette("recebimento/cartao/transacoes/sem_filtro") do
+        resposta = subject.transacoes
+        expect(resposta).to be_an(Array)
+        expect(resposta.first).to be_an(OpenStruct)
+      end
+    end
+
+    context "when TODO: fazer um teste para cada grupo de filtros (sem filtro, efetivado, por data, paginado) e outro passando vários juntos" do
+      it "TODO" do
+        VCR.use_cassette("recebimento/cartao/transacoes/x") do
+          pending("TODO")
+          fail
+        end
+      end
+    end
+  end
 end
